Update file.go
Note: this change does not impact the ability for users to use ReadFile function as they are currently using, is backwards compatible and does not change the core functionality of the function. 

The change recognizes if there as been an error during the file closing process and logs the error while the implementation of the function remains exactly the same.

Why this change:
-> Logging the error on f.Close is important for many applications.
-> The programmer has a right to know, if there was an error.
-> Errors should not be lost forever, a log gives access to errors if someone wants to see them.
-> at this stage, we should do at least the logging of the error and improve the ReadFile function to not have blind spots for security and efficiency, during audits.

I pray, kindly approve this commit.